refactor: extension 삭제 후 loadImage함수 DetailMovieViewController로 이동
